requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.4
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.18
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.1
decorator==5.1.1
Deprecated==1.2.14
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.104.1
filelock==3.13.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Dance==7.0.0
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.2
google-ai-generativelanguage==0.3.2
google-api-core==2.12.0
google-auth==2.23.2
google-generativeai==0.2.0
googleapis-common-protos==1.60.0
gpt4all==1.0.12
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipython==8.16.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kubernetes==28.1.0
langchain==0.0.304
langsmith==0.0.41
llama-index==0.8.36
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.8.1
numexpr==2.8.7
numpy==1.26.0
oauthlib==3.2.2
onnxruntime==1.16.3
openai==0.28.1
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
overrides==7.4.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
posthog==3.0.2
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.3
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pypdf==3.16.2
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.15.0
tomli==2.0.1
tqdm==4.66.1
traitlets==5.10.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
URLObject==2.4.3
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.7
websocket-client==1.6.4
websockets==12.0
Werkzeug==2.3.7
wrapt==1.16.0
yarl==1.9.2
zipp==3.17.0
